% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-shiny.R
\name{div_validation_error}
\alias{div_validation_error}
\title{Shiny validation error 'div'}
\usage{
div_validation_error(..., class = NULL)
}
\arguments{
\item{...}{passed to \code{div}}

\item{class}{'HTML' class}
}
\value{
A 'HTML' tag.
}
\description{
Used within shiny render functions to display shiny error
messages without using \code{\link[shiny]{validate}}.
}
